+++ packages/components/atoms/f-form-field/test/visual/f-form-field.visual.spec.js
@@ -1,17 +1,10 @@
-const FormField = require('../../test-utils/component-objects/f-form-field--visual.component');
+import FormField from '../../test-utils/component-objects/f-form-field--visual.component';
 
 describe('f-form-field visual tests', () => {
-    let formField;
-
-    beforeEach(async () => {
-        // Arrange
-        formField = new FormField();
-    });
-
     describe('default state', () => {
         it('should display all fields in the default state', async () => {
             // Act
-            await formField.load();
+            await FormField.load();
 
             // Assert
             await browser.percyScreenshot('f-form-field - Base State', 'desktop');
@@ -21,7 +14,7 @@ describe('f-form-field visual tests', () => {
     describe('disabled state', () => {
         it('should display all fields in a disabled state', async () => {
             // Act
-            await formField.load({ isDisabled: true });
+            await FormField.load({ isDisabled: true });
 
             // Assert
             await browser.percyScreenshot('f-form-field - Disabled State', 'desktop');
@@ -31,7 +24,7 @@ describe('f-form-field visual tests', () => {
     describe('errored state', () => {
         it('should display all fields in an errored state', async () => {
             // Act
-            await formField.load({ hasError: true });
+            await FormField.load({ hasError: true });
 
             // Assert
             await browser.percyScreenshot('f-form-field - Errored State', 'desktop');
@@ -41,7 +34,7 @@ describe('f-form-field visual tests', () => {
     describe('none required state', () => {
         it('should display all fields in an none required state', async () => {
             // Act
-            await formField.load({ isRequired: false });
+            await FormField.load({ isRequired: false });
 
             // Assert
             await browser.percyScreenshot('f-form-field - None required State', 'desktop');