+++ packages/components/atoms/f-image-tile/test/component/f-image-tile.component.spec.js
@@ -1,19 +1,17 @@
-const ImageTile = require('../../test-utils/component-objects/f-image-tile.component');
-
-const imageTile = new ImageTile();
+import ImageTile from '../../test-utils/component-objects/f-image-tile.component';
 
 describe('f-image-tile component tests', () => {
     beforeEach(async () => {
-        await imageTile.load();
+        await ImageTile.load();
     });
 
     it('should display the f-image-tile component', async () => {
         // Assert
-        await expect(await imageTile.isComponentDisplayed()).toBe(true);
+        await expect(await ImageTile.isComponentDisplayed()).toBe(true);
     });
 
     it('should check that the f-image-tile component is clickable', async () => {
         // Assert
-        await expect(await imageTile.isComponentClickable()).toBe(true);
+        await expect(await ImageTile.isComponentClickable()).toBe(true);
     });
 });