+++ packages/components/atoms/f-card/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.1.0
+------------------------------
+*July 4, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0 to use new Axe Core implementation.
+- Accessibility tests to be async.
+- Specs to ES6 syntax
+- Babel config to allow for async syntax in tests.
+
 v4.0.0
 ------------------------------
 *June 15, 2022*